requirements.txt
requests==2.31.0
pandas==2.0.3
google-cloud-bigquery==3.11.3
google-cloud-storage==2.10.0
gunicorn==21.2.0